media-source: move constants from AudioPlayer
